Ignores resource if it has a protocol or has no path
If the resource matched has a protocol it is not a path to CDN-ify. This
happens with links to about: or mailto: protocols.
Likewise if the resource has no path after being converted to a resourceUrl,
then it should be ignored as well (if the path is `undefined` then some
functions like `path.join` will fail).